backend/requirements.txt
langchain-community
openai
langserve
faiss-cpu
fastapi
uvicorn
beautifulsoup4
tiktoken
PyMuPDF